Fix comment and fix size of optimizer_backedge_threshold

+++ Include/internal/pycore_interp.h
@@ -239,8 +239,8 @@ struct _is {
     _PyOptimizerObject *optimizer;
     _PyExecutorObject *executor_list_head;
 
-    /* These three values are shifted and offset to speed up check in JUMP_BACKWARD */
-    uint32_t optimizer_backedge_threshold;
+    /* These values are shifted and offset to speed up check in JUMP_BACKWARD */
+    uint16_t optimizer_backedge_threshold;
     uint16_t optimizer_side_threshold;
 
     _rare_events rare_events;
